use node to run qunit unit tests
