+++ .eslintrc.json
@@ -1,0 +1,95 @@
+{
+	"root": true,
+	"parserOptions": {
+		"ecmaVersion": 6
+	},
+	"extends": [
+		"wikimedia/client-es5",
+		"wikimedia/jquery",
+		"wikimedia/mediawiki"
+	],
+	"env": {
+		"commonjs": true
+	},
+	"rules": {
+		"camelcase": "off",
+		"comma-dangle": "off",
+		"comma-spacing": "off",
+		"computed-property-spacing": "off",
+		"curly": "off",
+		"dot-notation": "off",
+		"eqeqeq": "off",
+		"func-call-spacing": "off",
+		"indent": "off",
+		"key-spacing": "off",
+		"keyword-spacing": "off",
+		"max-len": "off",
+		"max-statements-per-line": "off",
+		"new-cap": "off",
+		"no-alert": "off",
+		"no-extra-semi": "off",
+		"no-implicit-coercion": "off",
+		"no-jquery/no-animate-toggle": "off",
+		"no-jquery/no-bind": "off",
+		"no-jquery/no-browser": "off",
+		"no-jquery/no-each-util": "off",
+		"no-jquery/no-event-shorthand": "off",
+		"no-jquery/no-fade": "off",
+		"no-jquery/no-global-selector": "off",
+		"no-jquery/no-grep": "off",
+		"no-jquery/no-in-array": "off",
+		"no-jquery/no-is-array": "off",
+		"no-jquery/no-is-numeric": "off",
+		"no-jquery/no-map-util": "off",
+		"no-jquery/no-parse-html-literal": "off",
+		"no-jquery/no-parse-json": "off",
+		"no-jquery/no-ready-shorthand": "off",
+		"no-jquery/no-size": "off",
+		"no-jquery/no-slide": "off",
+		"no-jquery/no-trim": "off",
+		"no-jquery/no-unbind": "off",
+		"no-jquery/no-unique": "off",
+		"no-mixed-spaces-and-tabs": "off",
+		"no-multiple-empty-lines": "off",
+		"no-multi-spaces": "off",
+		"no-octal-escape": "off",
+		"no-prototype-builtins": "off",
+		"no-redeclare": "off",
+		"no-restricted-syntax": "off",
+		"no-shadow-restricted-names": "off",
+		"no-tabs": "off",
+		"no-throw-literal": "off",
+		"no-undef": "off",
+		"no-underscore-dangle": "off",
+		"no-unneeded-ternary": "off",
+		"no-unreachable": "off",
+		"no-unused-vars": "off",
+		"no-use-before-define": "off",
+		"no-useless-concat": "off",
+		"no-useless-escape": "off",
+		"no-whitespace-before-property": "off",
+		"object-curly-spacing": "off",
+		"operator-linebreak": "off",
+		"quote-props": "off",
+		"quotes": "off",
+		"semi": "off",
+		"semi-spacing": "off",
+		"space-before-blocks": "off",
+		"space-before-function-paren": "off",
+		"spaced-comment": "off",
+		"space-infix-ops": "off",
+		"space-in-parens": "off",
+		"space-unary-ops": "off",
+		"valid-jsdoc": "off",
+		"wrap-iife": "off",
+		"yoda": "off",
+		"no-jquery/no-class-state": "warn",
+		"no-jquery/no-sizzle": "warn",
+		"no-jquery/variable-pattern": "warn",
+		"no-shadow": "warn",
+		"no-unreachable-loop": "warn",
+		"mediawiki/msg-doc": "warn",
+		"compat/compat": "warn",
+		"block-scoped-var": "warn"
+	}
+}
