+++ codecov.yml
@@ -1,0 +1,2 @@
+fixes:
+  - "/var/www/html/extensions/SemanticFormsSelect/::"
